swapped AF handling to managed

+++ include/socklynx/endpoint.h
@@ -47,7 +47,7 @@ typedef union sl_endpoint_u {
 
 
 SL_INLINE_DECL uint16_t sl_endpoint_af_get(sl_endpoint_t *endpoint);
-SL_INLINE_DECL uint16_t sl_endpoint_af_set(sl_endpoint_t *endpoint, int32_t af);
+SL_INLINE_DECL int sl_endpoint_af_set(sl_endpoint_t *endpoint, uint16_t af);
 SL_INLINE_DECL struct sockaddr *sl_endpoint_addr_get(sl_endpoint_t *endpoint);
 SL_INLINE_DECL bool sl_endpoint_is_ipv4(sl_endpoint_t *endpoint);
 SL_INLINE_DECL bool sl_endpoint_is_ipv6(sl_endpoint_t *endpoint);
